Refactor the notifications to cache and reuse results from the unread-notifs checks
